# public_html/Swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Author Api",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.karmacraft.in",
            "description": "this is Local Server"
        }
    ],
    "paths": {
...
        "/api/addproject": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Add a new project",
                "operationId": "7b0864f8ccaedd7a35b3b7ed717700a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_name": {
                                        "type": "string",
                                        "example": "Project A"
                                    },
                                    "Project_buget": {
                                        "type": "number",
                                        "example": 10000
                                    },
                                    "project_description": {
                                        "type": "string",
                                        "example": "Description of Project A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Project Added Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ststus": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fail to Register Due to Incomplete Fields"
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "project_name": "The project_name field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Cant add project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cant Add Project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getproject": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get projects for the authenticated user",
                "operationId": "547903cb2a26438f37f1b452b93f6835",
                "responses": {
                    "200": {
                        "description": "Projects retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "projects": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 123
                                                    },
                                                    "project_name": {
                                                        "type": "string",
                                                        "example": "Project A"
                                                    },
                                                    "Project_buget": {
                                                        "type": "number",
                                                        "example": 10000
                                                    },
                                                    "project_description": {
                                                        "type": "string",
                                                        "example": "Description of Project A"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-26T10:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-26T10:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No projects found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "No Projects Founds"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}